add moderation controller
