+++ server/api/controllers/screenshotController.js
@@ -11,12 +11,10 @@ module.exports = {
   getfromId,
   getUnsolvedScreenshot,
   getLastAddedScreenshot,
-  getNonModeratedScreenshots,
   removeOwnScreenshot,
   tryProposal,
   uploadScreenshot,
   addScreenshot,
-  moderateScreenshot,
 };
 
 async function getfromId(req) {
@@ -86,26 +84,6 @@ async function getUnsolvedScreenshot(req) {
 async function getLastAddedScreenshot(req) {
   const screenshotId = await screenshotManager.getLastAdded();
   return getfromId({ ...req, body: { ...req.body, id: screenshotId } });
-}
-
-async function getNonModeratedScreenshots(req) {
-  const { user } = req;
-  if (!user) {
-    return bluebird.reject({
-      status: 401,
-      code: 'MUST_BE_IDENTIFIED',
-      message: "User must be identified to proove that he/she's a moderator.",
-    });
-  }
-  if (!user.canModerateScreenshots) {
-    return bluebird.reject({
-      status: 403,
-      code: 'CANNOT_MODERATE_SCREENSHOTS',
-      message: 'User has no right to moderate screenshots.',
-    });
-  }
-  const screenshots = await screenshotManager.getNonModeratedScreenshots();
-  return screenshots.map(addImageUrlFromPath);
 }
 
 async function removeOwnScreenshot(req) {
@@ -209,26 +187,6 @@ async function addScreenshot(req) {
   return screenshotManager.create(screenshot);
 }
 
-async function moderateScreenshot(req) {
-  const { user } = req;
-  const { screenshotId, approve } = req.body;
-  if (!user) {
-    return bluebird.reject({
-      status: 401,
-      code: 'MUST_BE_IDENTIFIED',
-      message: 'User must be identified to approve screenshots.',
-    });
-  }
-  return screenshotManager.moderateScreenshot({ screenshotId, user, approve });
-}
-
 function getUploadedImageLocalPath(imageName) {
   return `${__dirname}/../../uploads/${imageName}`;
-}
-
-function addImageUrlFromPath(screenshot) {
-  return {
-    ...screenshot,
-    imageUrl: cloudinaryService.pathToUrl(screenshot.imagePath),
-  };
 }
