+++ server/api/managers/screenshotManager.js
@@ -7,11 +7,9 @@ module.exports = {
   getFromId,
   getLastAdded,
   getUnsolved,
-  getNonModeratedScreenshots,
   deleteUserScreenshot,
   testProposal,
   markScreenshotAsResolved,
-  moderateScreenshot,
 };
 
 async function create(screenshotToCreate) {
@@ -160,23 +158,6 @@ async function getUnsolved({ userId, exclude }) {
   return screenshots[0];
 }
 
-async function getNonModeratedScreenshots() {
-  const results = await db.Screenshot.findAll({
-    attributes: ['id', 'gameCanonicalName', 'year', 'imagePath', 'createdAt'],
-    where: { approvalStatus: 0 },
-    limit: 100,
-    order: [['createdAt', 'ASC']],
-  });
-  return results.map(res => ({
-    id: res.id,
-    name: res.gameCanonicalName,
-    year: res.year || null,
-    createdAt: res.createdAt,
-    imagePath: res.imagePath,
-    awaitingApproval: true,
-  }));
-}
-
 async function deleteUserScreenshot({ userId, screenshotId }) {
   const screenshot = await db.Screenshot.findOne({
     attributes: ['id'],
@@ -264,38 +245,6 @@ async function markScreenshotAsResolved({ screenshotId, userId }) {
     user.addSolvedScreenshot(solvedScreenshot),
     screenshot.addSolvedScreenshot(solvedScreenshot),
     user.increment('solvedScreenshots'),
-  ]);
-}
-
-async function moderateScreenshot({ screenshotId, user, approve }) {
-  const [moderator, screenshot] = await Promise.all([
-    db.User.findById(user.id),
-    db.Screenshot.findById(screenshotId),
-  ]);
-  console.log({ approve, approvalStatus: screenshot.approvalStatus });
-  if (!moderator) {
-    throw new Error('Moderator not found');
-  }
-  if (!screenshot) {
-    throw new Error('Screenshot not found');
-  }
-  if (approve && screenshot.approvalStatus === 1) {
-    return;
-  }
-  if (!approve && screenshot.approvalStatus === -1) {
-    return;
-  }
-  const shouldDecrement = !approve && screenshot.approvalStatus === 1;
-  console.log('increment', approve);
-  console.log('decrement', shouldDecrement);
-  const poster = await db.User.findById(screenshot.UserId);
-  await Promise.all([
-    screenshot.update({
-      approvalStatus: approve ? 1 : -1,
-      moderatedBy: moderator.id,
-    }),
-    shouldDecrement ? poster.decrement('addedScreenshots') : null,
-    approve ? poster.increment('addedScreenshots') : null,
   ]);
 }
