+++ src/services/screenshotService.js
@@ -6,10 +6,8 @@ export default {
   getFromId,
   getLast,
   getUnsolved,
-  getNonModerated,
   guess,
   removeOwn,
-  moderate,
 };
 
 function uploadImage(file, onUploadProgress) {
@@ -36,10 +34,6 @@ function getUnsolved(exclude) {
   return api.post('/screenshot/unsolved', { exclude });
 }
 
-function getNonModerated() {
-  return api.post('/screenshot/non-moderated');
-}
-
 function guess(screenshotId, proposal) {
   return api.post('/screenshot/guess', {
     screenshotId,
@@ -50,12 +44,5 @@ function guess(screenshotId, proposal) {
 function removeOwn(screenshotId) {
   return api.post('/screenshot/remove-own', {
     screenshotId,
-  });
-}
-
-function moderate({ screenshotId, approve }) {
-  return api.post('/screenshot/moderate', {
-    screenshotId,
-    approve,
   });
 }
